+++ gems/aws-sdk-s3/lib/aws-sdk-s3/encryption/default_cipher_provider.rb
@@ -12,7 +12,7 @@ module Aws
 
         # @return [Array<Hash,Cipher>] Creates an returns a new encryption
         #   envelope and encryption cipher.
-        def encryption_cipher
+        def encryption_cipher(options = {})
           cipher = Utils.aes_encryption_cipher(:CBC)
           envelope = {
             'x-amz-key' => encode64(encrypt(envelope_key(cipher))),
@@ -24,7 +24,7 @@ module Aws
 
         # @return [Cipher] Given an encryption envelope, returns a
         #   decryption cipher.
-        def decryption_cipher(envelope)
+        def decryption_cipher(envelope, options = {})
           master_key = @key_provider.key_for(envelope['x-amz-matdesc'])
           key = Utils.decrypt(master_key, decode64(envelope['x-amz-key']))
           iv = decode64(envelope['x-amz-iv'])